refactor: :recycle: オーバーレイローディングとトースト表示の処理を presentation 層に移動
