+++ lib/core/router/app_router.dart
@@ -16,7 +16,7 @@ import '../../feature/introduction/presentation/introduction_page.dart';
 import '../../feature/setting/presentation/license_page.dart';
 import '../../feature/setting/presentation/muted_user_list_page.dart';
 import '../../feature/setting/presentation/setting_page.dart';
-import '../../feature/user_follow/presentation/page/following_and_follower_list/following_and_follower_list_page.dart';
+import '../../feature/user_follow/presentation/page/following_and_follower_list_page.dart';
 import '../../feature/user_profile/presentation/page/edit_profile/edit_profile_page.dart';
 import '../../feature/user_profile/presentation/page/like_user_page.dart/like_user_page.dart';
 import '../../feature/user_profile/presentation/page/profile/profile_page.dart';
