+++ lib/feature/user_profile/presentation/component/profile_list.dart
@@ -2,9 +2,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../core/common_widget/button/follow_or_unfollow_button.dart';
 import '../../../../core/common_widget/infinity_scroll_widget.dart';
 import '../../../auth/application/auth_state.dart';
+import '../../../user_follow/presentation/component/follow_or_unfollow_button.dart';
 import '../../application/user_id_list_state.dart';
 import '../../util/user_list_type.dart';
 import 'profile_tile.dart';
