+++ lib/feature/user_profile/presentation/page/profile/profile_widget.dart
@@ -4,12 +4,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../../../core/common_widget/avatar_network_image_widget.dart';
-import '../../../../../core/common_widget/button/follow_or_unfollow_button.dart';
 import '../../../../../core/common_widget/button/outlined_button.dart';
 import '../../../../../core/common_widget/error_and_retry_widget.dart';
 import '../../../../../core/router/app_router.dart';
 import '../../../../../util/logger.dart';
 import '../../../../auth/application/auth_state.dart';
+import '../../../../user_follow/presentation/component/follow_or_unfollow_button.dart';
 import '../../../../user_follow/presentation/component/following_and_follower_count_widget.dart';
 import '../../../application/user_profile_state.dart';
 import 'profile_widget_shimmer.dart';
